feat: adds scroll-to-top button for Wallet40 nav
adds accessible label to scroll button
