+++ .changeset/little-games-hide.md
@@ -1,0 +1,5 @@
+---
+"ledger-live-desktop": minor
+---
+
+feat(lwd): replace scroll to top component
