+++ apps/ledger-live-desktop/src/mvvm/components/Page/components/index.ts
@@ -1,3 +1,4 @@
 export { ClassicLayout, getPagePaddingLeft, getPagePaddingRight } from "./ClassicLayout";
 export { Wallet40Layout } from "./Wallet40Layout";
+export { LegacyScrollUpButton } from "./LegacyScrollUpButton";
 export { ScrollUpButton } from "./ScrollUpButton";
